requirements.txt
requests
dotenv
tzdata
python-telegram-bot